// libs/types/follow/follow.ts
import { MeLiked } from "../like/like";
import { Member } from "../members/member";
import { MetaCounter } from "../doctors/doctor";

export interface MeFollowed {
  followingId: string;
  followerId: string;
  myFollowing: boolean;
}

export interface Follower {
  _id: string;
  followingId: string;
  followerId: string;
  createdAt: Date;
  updatedAt: Date;

  /** from aggregation **/
  meLiked?: MeLiked;
  meFollowed?: MeFollowed;
  followerData?: Member;
}

export interface Following {
  _id: string;
  followingId: string;
  followerId: string;
  createdAt: Date;
  updatedAt: Date;

  /** from aggregation **/
  meLiked?: MeLiked;
  meFollowed?: MeFollowed;
  followingData?: Member;
}

export interface Followings {
  list: Following[];
  metaCounter: MetaCounter[];
}

export interface Followers {
  list: Follower[];
  metaCounter: MetaCounter[];
}